+++ R/lms.R
@@ -1,10 +1,15 @@
 #' Adjusted summary for linear regression when based on replicated dataset
-#' @description adjusting lm object results according to original number of degree of freedom.
-#' The standard errors, t statistics and p values have to be adjusted because of replicated observations.
+#' @description adjusting lm object results according to original number of
+#' degree of freedom.
+#' The standard errors, t statistics and p values have to be adjusted because of
+#'  replicated observations.
 #' @param x lm object
-#' @param df_old integer number of d.f in original dataset. For bigger datasets `nrow` should be sufficient.
-#' @param df_new integer number of d.f in dataset with replicated rows, Default: x$df.residual
-#' @return data.frame with additional columns over a regular summary.lm output, like correct and statistics adjusted by it.
+#' @param df_old integer number of d.f in original dataset. For bigger datasets
+#' `nrow` should be sufficient.
+#' @param df_new integer number of d.f in dataset with replicated rows,
+#' Default: x$df.residual
+#' @return data.frame with additional columns over a regular summary.lm output,
+#' like correct and statistics adjusted by it.
 #' @importFrom stats pt
 #' @details The size of the correction is equal to sqrt(df_new / df_old).
 #' Where standard errors are multiplied and t statistics divided by it.
@@ -18,7 +23,12 @@
 #' occup_new <- occup_small[occup_small$year == 2010, ]
 #'
 #' occup_2 <- cat2cat(
-#'   data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
+#'   data = list(
+#'     old = occup_old,
+#'     new = occup_new,
+#'     cat_var = "code",
+#'     time_var = "year"
+#'   ),
 #'   mappings = list(trans = trans, direction = "backward"),
 #'   ml = list(
 #'     data = occup_new,
@@ -31,7 +41,9 @@
 #'
 #' # Regression
 #' # we have to adjust size of std as we artificialy enlarge degrees of freedom
-#' lms <- lm(I(log(salary)) ~ age + sex + factor(edu) + parttime + exp, occup_2$old,
+#' lms <- lm(
+#'   formula = I(log(salary)) ~ age + sex + factor(edu) + parttime + exp,
+#'   data = occup_2$old,
 #'   weights = multiplier * wei_freq_c2c
 #' )
 #'